login modificaciones

+++ src/app/servicios/proyectos.service.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
@@ -9,21 +9,31 @@ export class ProyectosService {
 
   constructor(private httpClient: HttpClient) { }
   getDatos(url: string): Observable<any> {
-    return this.httpClient.get(url);
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.get(url,{headers:{'Authorization': sessionStorage['token']}});
+
+    // const headers = new HttpHeaders();
+    // headers.set('Authorization', sessionStorage.getItem('token'));
+    // return this.httpClient.get(url, {headers:headers})
   }
 
   saveOrUpdate(url:string, objeto:any):Observable<any>{
-    console.log("objeto:"+objeto)
-    return this.httpClient.post(url,objeto)
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.post(url, {headers:{'Authorization': sessionStorage['token']}});
   }
 
   delete(url:string):Observable<any>{
-    return this.httpClient.delete(url);
+    const headers = new HttpHeaders();
+    headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.delete(url, {headers: {'Authorization': sessionStorage['token']}});
   }
 
-  ifExist(url:string):Observable<any>{
-  
-    return this.httpClient.post(url,"");
+  patch(url:string, user:any):Observable<any>{
+  const headers = new HttpHeaders;
+  headers.set('Authorization', sessionStorage['token']);
+    return this.httpClient.patch(url, 'user', {headers:{'Authorization': sessionStorage['token']}});
 
   }
